LayoutTests/platform/gtk/accessibility/media-element-expected.txt

Dump <video> element controller accessibility object tree at 'canplaythrough' event.

+++++++++++++++++++++++++++++++++++

State at 'canplaythrough' event:

    description: AXDescription: video element controller
    role: AXRole: AXToolbar


        description: AXDescription: video element controller
        role: AXRole: AXToolbar


            description: AXDescription: play
            role: AXRole: AXButton


            description: AXDescription: 
            role: AXRole: AXSlider


            description: AXDescription: enter fullscreen
            role: AXRole: AXButton


            description: AXDescription: mute
            role: AXRole: AXButton
